Prevent cleanup from running twice

+++ test/integration/targets/foreman_inventory/runme.sh
@@ -9,13 +9,17 @@ export FOREMAN_HOST="${FOREMAN_HOST:-localhost}"
 export FOREMAN_PORT="${FOREMAN_PORT:-8080}"
 FOREMAN_CONFIG=test-config.foreman.yaml
 
+_is_clean=  # flag for checking whether cleanup has already fired
+
 function _cleanup() {
+    [[ -n "$_is_clean" ]] && return  # don't double-clean
     echo Cleanup: removing $FOREMAN_CONFIG...
     rm -vf "$FOREMAN_CONFIG"
     unset ANSIBLE_CONFIG
     unset FOREMAN_HOST
     unset FOREMAN_PORT
     unset FOREMAN_CONFIG
+    _is_clean=1
 }
 trap _cleanup INT TERM EXIT
